add logout in traveller UI

+++ src/app/user/user/user.component.ts
@@ -102,9 +102,14 @@ getActiveTravellers()
 
    logOut()
   {
+    if (confirm("Sure , you want to log out") == true) {
     sessionStorage.clear() ;
     this.toast.success({detail:"SUCCESS",summary:'Log Out Successfully',duration:2000});
     this.router.navigate(["login"]);
+    }
+    else{
+    }
+    
   }
